# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CinePrime",
    "description": "CinePrime is a movie booking platform that allows users to book movies and watch them in the cinema.",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/movie/add": {
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "Add New Movie",
        "description": "Add New Movie",
        "operationId": "addNewMovie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "director": {
                    "type": "string",
                    "example": "Kari Pollard"
                  },
                  "duration": {
                    "type": "number",
                    "example": 106
                  },
                  "genre": {
                    "type": "string",
                    "example": "rest"
                  },
                  "overview": {
                    "type": "string",
                    "example": "Difference of describe time peace goal. Those concern pretty force most which bring.\nBlue leg allow company series. Season bank individual guy west per local. Heart trade produce economy."
                  },
                  "rating": {
                    "type": "string",
                    "example": "PG"
                  },
                  "release_date": {
                    "type": "string",
                    "example": "2022-10-26"
                  },
                  "title": {
                    "type": "string",
                    "example": "Adaptive coherent support"
                  }
                }
              },
              "examples": {
                "Add New Movie": {
                  "value": {
                    "director": "Kari Pollard",
                    "duration": 106,
                    "genre": "rest",
                    "overview": "Difference of describe time peace goal. Those concern pretty force most which bring.\nBlue leg allow company series. Season bank individual guy west per local. Heart trade produce economy.",
                    "rating": "PG",
                    "release_date": "2022-10-26",
                    "title": "Adaptive coherent support"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201 Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 201
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 48
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Created"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                },
                "examples": {
                  "201 Created": {
                    "value": {
                      "code": 201,
                      "data": {
                        "id": 48
                      },
                      "message": "Created",
                      "success": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}